tests/wpt/web-platform-tests/css/css-text/animations/letter-spacing-interpolation.html

<!DOCTYPE html>
<meta charset="UTF-8">
<link rel="help" href="https://drafts.csswg.org/css-text-3/#letter-spacing-property">
<meta name="test" content="letter-spacing supports animation by computed value type">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<style>
.parent {
  letter-spacing: 2px;
}
.target {
  font: 10px sans-serif;
  font-weight: bold;
  letter-spacing: 30px;
}
.expected {
  color: green;
  margin-right: 30px;
}
</style>

<body>
  <template id="target-template">TT</template>
</body>

<script>
test_interpolation({
  property: 'letter-spacing',
  from: neutralKeyframe,
  to: '20px',
}, [
  {at: -0.3, expect: '33px'},
  {at: 0, expect: '30px'},
  {at: 0.3, expect: '27px'},
  {at: 0.6, expect: '24px'},
  {at: 1, expect: '20px'},
  {at: 1.5, expect: '15px'},
]);

test_interpolation({
  property: 'letter-spacing',
  from: 'initial',
  to: '20px',
}, [
  {at: -0.3, expect: '-6px'},
  {at: 0, expect: 'normal'},
  {at: 0.3, expect: '6px'},
  {at: 0.6, expect: '12px'},
  {at: 1, expect: '20px'},
  {at: 1.5, expect: '30px'},
]);

test_interpolation({
  property: 'letter-spacing',
  from: 'inherit',
  to: '20px',
}, [
  {at: -0.3, expect: '-3.4px'},
  {at: 0, expect: '2px'},
  {at: 0.3, expect: '7.4px'},
  {at: 0.6, expect: '12.8px'},
  {at: 1, expect: '20px'},
  {at: 1.5, expect: '29px'},
]);

test_interpolation({
  property: 'letter-spacing',
  from: 'unset',
  to: '20px',
}, [
  {at: -0.3, expect: '-3.4px'},
  {at: 0, expect: '2px'},
  {at: 0.3, expect: '7.4px'},
  {at: 0.6, expect: '12.8px'},
  {at: 1, expect: '20px'},
  {at: 1.5, expect: '29px'},
]);

test_interpolation({
  property: 'letter-spacing',
  from: '-10px',
  to: '10px',
}, [
  {at: -0.3, expect: '-16px'},
  {at: 0, expect: '-10px'},
  {at: 0.3, expect: '-4px'},
  {at: 0.6, expect: '2px'},
  {at: 1, expect: '10px'},
  {at: 1.5, expect: '20px'},
]);

test_interpolation({
  property: 'letter-spacing',
  from: 'normal',
  to: '10px',
}, [
  {at: -0.3, expect: '-3px'},
  {at: 0, expect: 'normal'},
  {at: 0.3, expect: '3px'},
  {at: 0.6, expect: '6px'},
  {at: 1, expect: '10px'},
  {at: 1.5, expect: '15px'},
]);
</script>
